This is the final version of the site before the showcase on 4/4/23
I changed a minor details of formatting on most pages and added some
logic which is commented out for now that when implemented will make the
pages more robust. The only reason these arent fully implemented right
now is because they are nessacary for the prototype showcase and they
are better off implemented with the further imporvments we'll be adding
down the line as we incorporate more of the user stories.