Commiting 75% of the way there
Create,Read,Update(updating the price) features are complete as of this
commit, the only thing left is Delete which would be an easy
implemention except for the fact that device id is a foreign key in the
orderline table so it doesnt let you just rip it out, hence I am
assuming I just have to implement queries in the delete method that
search for records containing the id in the orderline table and delete
them first, or even find a way to detach those numbers from the
references objects in some way and just leave them in there but not
referenceing anything, although for the purposes of presenting this as a
project I dont think I need to go that deep.

Only other issue is the testing files just broke completly on me, I am
most likley going to have to look at this and fix it otehrwise itll
break main when it gets pushed but it's 3:11am and I'm not doing it now.

Past that I will go over my nav bars for the pages I've worked on and
just clean them up and make everything like generally presentable in the
areas that I've edited