Completed device
Just did the finishing touches on device adding/deleting/updating

Made it so that the user is redirected to different pages when these
events occur and that the topnav bar makes sense on all these pages.

Also added input validation to the update section and the add device
section so they cant add blank devices, all of them have to have at
least a name, price and soh value.

If there are any other errors found in testing I can fix them tomorrow
during the day